granthik modulized/detailed/GRANTHIK_Final_Modular_Backend/backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
bcrypt==4.0.1
pydantic>=2.4.2
pydantic-settings>=2.0.0
sqlalchemy>=2.0.22
psycopg2-binary>=2.9.9
alembic>=1.12.0
chromadb>=0.4.22
langchain>=0.2.0
langchain-community>=0.0.6
langchain-ollama>=0.1.0
langchain-chroma>=0.1.0
langchain-core>=0.1.0
sentence-transformers>=2.2.2
llama-index>=0.9.0
llama-index-llms-ollama>=0.1.0
llama-index-embeddings-huggingface>=0.1.0
requests>=2.31.0
python-dotenv>=1.0.0
pytest>=7.4.2
httpx>=0.25.0
python-docx>=0.8.11
torch>=2.0.0
transformers>=4.30.0
psutil>=5.9.0  # For memory usage monitoring
reportlab>=4.0.0  # For PDF generation
markdown>=3.4.0  # For markdown processing
weasyprint>=59.0  # For HTML to PDF conversion